Implementing message system library Referenced the lib in report-sink and notifications-rules-engine Created common code for plugin and its configuration
