+++ libs/message-system/src/main/kotlin/gov.cdc.ocio.messagesystem/MessageSystemProcessor.kt
@@ -1,0 +1,7 @@
+package gov.cdc.ocio.messagesystem
+
+import com.azure.messaging.servicebus.ServiceBusReceivedMessageContext
+
+interface MessageSystemProcessor {
+    fun processMessage(context: ServiceBusReceivedMessageContext)
+}
