+++ libs/message-system/src/main/kotlin/gov.cdc.ocio.messagesystem/config/RabbitMQServiceConfiguration.kt
@@ -1,4 +1,4 @@
-package gov.cdc.ocio.messagesystem.plugins
+package gov.cdc.ocio.messagesystem.config
 
 import com.rabbitmq.client.*
 import io.ktor.server.config.*
