+++ libs/message-system/src/main/kotlin/gov.cdc.ocio.messagesystem/health/HealthCheckRabbitMQ.kt
@@ -3,7 +3,7 @@ package gov.cdc.ocio.messagesystem.health
 import com.rabbitmq.client.*
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties
 import com.rabbitmq.client.Channel
-import gov.cdc.ocio.messagesystem.plugins.RabbitMQServiceConfiguration
+import gov.cdc.ocio.messagesystem.config.RabbitMQServiceConfiguration
 import gov.cdc.ocio.types.health.HealthCheckResult
 import gov.cdc.ocio.types.health.HealthCheckSystem
 import gov.cdc.ocio.types.health.HealthStatusType
